flush source stream on return
